belajar-fastapi/requirements.txt
fastapi==0.111.0
peewee==3.17.5
psycopg2==2.9.9
PyJWT==2.8.0
PyJWT==2.8.0
python_bcrypt==0.3.2
uvicorn==0.30.0
